[Feature]FastAPI 라우팅 분리
